Switch back to more performant and portable Marvin implementation
